requirements.txt
streamlit
google-generativeai
pypdf
protobuf<=3.20.3